# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Discovery-маркетплейс специалистов: каталог, поиск, лиды, отзывы.",
        "title": "marketpclce API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/categories/stats": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "search"
                ],
                "summary": "Кол-во специалистов в каждой категории",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_search.categoryStatsResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/internal_search.errorResponse"
                        }
                    }
                }
            }
        },
...
        "/reviews": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Оставляет отзыв на специалиста. Если указан lead_id, то клиент должен быть автором лида, а target — принятым получателем.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reviews"
                ],
                "summary": "Создать отзыв",
                "parameters": [
                    {
                        "description": "review payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_reviews.createReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_reviews.Review"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_reviews.errResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_reviews.errResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/internal_reviews.errResponse"
                        }
                    }
                }
            }
        },
        "/reviews/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Удаляет отзыв. Только автор отзыва.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reviews"
                ],
                "summary": "Удалить отзыв",
                "parameters": [
                    {
                        "type": "string",
                        "description": "review id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "no content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_reviews.errResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/internal_reviews.errResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_reviews.errResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Изменяет рейтинг и/или текст отзыва. Только автор отзыва.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reviews"
                ],
                "summary": "Изменить отзыв",
                "parameters": [
                    {
                        "type": "string",
                        "description": "review id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "patch payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_reviews.UpdateInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_reviews.Review"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_reviews.errResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_reviews.errResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/internal_reviews.errResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_reviews.errResponse"
                        }
                    }
                }
            }
        },
        "/search": {
            "get": {
                "description": "Возвращает hits + relaxed/similar если soft-фильтры не сошлись.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "search"
                ],
                "summary": "Поиск специалистов",
                "parameters": [
                    {
                        "type": "string",
                        "description": "search text",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "category codes (multi)",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "skill slugs (multi)",
                        "name": "skill",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "city",
                        "name": "city",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "rate min",
                        "name": "rate_min",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "rate max",
                        "name": "rate_max",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_search.Result"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/internal_search.errorResponse"
                        }
                    }
                }
            }
        },
...
        "/specialists/{id}/reviews": {
            "get": {
                "description": "Листит отзывы, отсортированные по дате убывания.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reviews"
                ],
                "summary": "Отзывы на специалиста (пагинация)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "specialist user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "default 20, max 100",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "default 0",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_reviews.listResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_reviews.errResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "internal_reviews.Review": {
            "type": "object",
            "properties": {
                "author_name": {
                    "type": "string"
                },
                "author_user_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lead_id": {
                    "type": "string"
                },
                "rating": {
                    "type": "integer"
                },
                "target_user_id": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "internal_reviews.UpdateInput": {
            "type": "object",
            "properties": {
                "rating": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "internal_reviews.createReq": {
            "type": "object",
            "properties": {
                "author_name": {
                    "type": "string"
                },
                "lead_id": {
                    "type": "string"
                },
                "rating": {
                    "type": "integer"
                },
                "target_user_id": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "internal_reviews.errResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "internal_reviews.listResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_reviews.Review"
                    }
                }
            }
        },
        "internal_search.CategoryCount": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "count": {
                    "type": "integer"
                }
            }
        },
        "internal_search.Facets": {
            "type": "object",
            "properties": {
                "categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_search.CategoryCount"
                    }
                }
            }
        },
        "internal_search.IndexDoc": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "bio": {
                    "type": "string"
                },
                "categories": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "city": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "display_name": {
                    "type": "string"
                },
                "is_published": {
                    "type": "boolean"
                },
                "primary_category": {
                    "type": "string"
                },
                "rate_max": {
                    "type": "integer"
                },
                "rate_min": {
                    "type": "integer"
                },
                "rating_avg": {
                    "type": "number"
                },
                "reviews_count": {
                    "type": "integer"
                },
                "skill_slugs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "skill_titles": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "internal_search.Result": {
            "type": "object",
            "properties": {
                "broadened": {
                    "type": "boolean"
                },
                "facets": {
                    "$ref": "#/definitions/internal_search.Facets"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_search.IndexDoc"
                    }
                },
                "relaxed": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "similar": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_search.IndexDoc"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "internal_search.categoryStatsResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_search.CategoryCount"
                    }
                }
            }
        },
        "internal_search.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}